Clean code [Hint below]
There is no interest to use itertools builtin function.
I tried an implementation which starts counting the number of repeats and remains
and use the zip() and islice() builtin function but the code is horrible to understand (maintain)
and there is no real performances gain to win this trade off.